requirements.txt
datasets
pytorch-crf
transformers